:rocket: Extract banner text component
